backend/requirements.txt
flask
flask-cors
spacy
ibm-watsonx-ai
python-dotenv
